core: Switch to CPU ID as parameter for arch_cpu_restore()
This avoids passing a pointer to the CPU-local mapping during shutdown
which will go away in the middle of this function. Rather specify the
current CPU ID so that the implementations can retrieve their own
references as needed. We cannot drop this parameter completely because
arch_cpu_restore() might be called before the CPU-local mapping was
fully established.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>