arm: Hide private per-cpu section during hypervisor operation
Turn off access to the private per-cpu section before the final
activation of the hypervisor. This prevents that one CPU can potentially
speculate about the private content of per-cpu sections of others.

We do not need to worry about the shutdown, though: The stack is
switched atomically with the MMU settings.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>